dispatch: make public interfaces LLP64 friendly
This attempts to replace the public uses of `unsigned long` with
`uintptr_t` and `long` with `intptr_t`.  The use of `long` and `unsigned
long` here was as a type suitable for matching pointer width.  However,
on LLP64 targets, this does not hold.  Replace them with `intptr_t`
which gives a proper pointer sized type.  This is done only for the
argument types and not for named types to ensure that the ABI is not
modified for C++ (where a template may be specialized on the named
type).

+++ dispatch/queue.h
@@ -454,7 +454,7 @@ typedef unsigned int dispatch_qos_class_t;
 API_AVAILABLE(macos(10.6), ios(4.0))
 DISPATCH_EXPORT DISPATCH_CONST DISPATCH_WARN_RESULT DISPATCH_NOTHROW
 dispatch_queue_t
-dispatch_get_global_queue(long identifier, unsigned long flags);
+dispatch_get_global_queue(intptr_t identifier, uintptr_t flags);
 
 /*!
  * @typedef dispatch_queue_attr_t
